PODS-1548 : updated test data files
